Finalizing html and css
me and Ian worked on these pages together.
In the front_end we added some icons to the header and changed some colours.
we added new pages to the framework and linked them with css.
a problem come up with the css, it wasn't linking properly, but after some conculting with together with luca, we solved the issue.
changed the localhost found in the config\App.php and changed the baseURL and default database in .env
changed the files from .html to .php and renaming them properly.