Updating the Html and Css
Me and Ian worked togethor to fix some of the html and css code to be responsive and pretty.
Connected some of the buttons, so that the navigation works between the pages.
Ian added and fixed images on certain pages that needed an image.
Fixed the forms textarea and inputs to be responsive with the page.
Also fixed some issues that poped up.